Switched to yabf

+++ example/cli.rs
@@ -20,18 +20,18 @@ fn main() -> Result<(), CenterlineError> {
     );
 
     let rejected_edges = centerline.rejected_edges().unwrap();
-    centerline.diagram().debug_print_all(|x:usize|!rejected_edges.bit(x as u64));
+    centerline.diagram().debug_print_all(|x:usize|!rejected_edges.bit(x));
 
     for (eid, e) in centerline
         .diagram()
         .edges()
         .iter()
         .enumerate()
-        .filter(|(i, _)| !rejected_edges.bit(*i as u64))
+        .filter(|(i, _)| !rejected_edges.bit(*i))
     {
         let twin = centerline.diagram().edge_get_twin(Some(e.get().get_id()));
-        let twin_i:Vec<bool> = twin.iter().map(|x| rejected_edges.bit(x.0 as u64)).collect();
-        print!("edge#{} {:?} {} {:?} {:?}", eid, e, rejected_edges.bit(eid as u64), twin, twin_i.first());
+        let twin_i:Vec<bool> = twin.iter().map(|x| rejected_edges.bit(x.0)).collect();
+        print!("edge#{} {:?} {} {:?} {:?}", eid, e, rejected_edges.bit(eid), twin, twin_i.first());
         let eid = Some(e.get().get_id());
         let v0 = centerline.diagram().edge_get_vertex0(eid).unwrap();
         let v0 = centerline.diagram().vertex_get(Some(v0)).unwrap().get();
